criacao do componente de visualizacao e prontuarios dos animais

+++ backend/veterinaria/src/main/java/br/com/softvet/veterinaria/model/repository/ProprietarioRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.softvet.veterinaria.model.entity.Proprietario;
 
-public interface ProprietarioRepository extends JpaRepository<Proprietario, Integer> {
+public interface ProprietarioRepository extends JpaRepository<Proprietario, Long> {
+	
 
 }
